refactor: README 변경, view 패키지 분리
