+++ src/main/java/README.md
@@ -1,0 +1,32 @@
+# step1
+## 기능 구현 목록
+1. 입력 View
+2. 0에서 9사이 랜덤 값 구하기
+4. 4이상일 때 전진
+5. 출력 View
+
+## 기능 요구사항
+* 초간단 자동차 경주 게임을 구현한다.
+* 주어진 횟수 동안 n대의 자동차는 전진 또는 멈출 수 있다.
+* 사용자는 몇 대의 자동차로 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
+* 전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4이상일 경우이다.
+* 자동차의 상태를 화면에 출력한다. 어느 시점에 출력할 것인지에 대한 제약은 없다.
+
+# step2
+## 기능 구현 목록
+* 자동차 이름 입력
+* 자동차 이름 쉼표 기준으로 구분
+* 우승자 선정
+* 실행 결과에 자동차 이름 출력
+* 우승자 출력
+
+## 기능 요구사항
+* 각 자동차에 이름을 부여할 수 있다. 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
+* 자동차 이름은 쉼표(,)를 기준으로 구분한다.
+* 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한명 이상일 수 있다.
+
+## step4 - 자동차 경주(리펙토링)
+## 리팩토링 요구사항
+* 핵심 비지니스 로직을 가지는 객체를 domain 패키지, UI 관련한 객체를 view 패키지에 구현한다.
+* MVC 패턴 기반으로 리팩토링해 view 패키지의 객체가 domain 패키지 객체에 의존할 수 있지만, domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현한다.
+* 테스트 가능한 부분과 테스트하기 힘든 부분을 분리해 테스트 가능한 부분에 대해서만 단위 테스트를 진행한다.
