--- src/main/java/step3/README.md
@@ -1,11 +1,0 @@
-## 기능구현목록
-* 자동차 이름 입력
-* 자동차 이름 쉼표 기준으로 구분
-* 우승자 선정
-* 실행 결과에 자동차 이름 출력
-* 우승자 출력
-
-## 기능 요구사항
-* 각 자동차에 이름을 부여할 수 있다. 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
-* 자동차 이름은 쉼표(,)를 기준으로 구분한다.
-* 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한명 이상일 수 있다.
